Delete dist/bundle.js and dist/bundle.js.map
These don't need to be committed. They're generated in the build
process, but don't need to be tracked by git.

.github/workflows/build-and-deploy.yml will perform the build,
and then publish the generated contents in the dist/ folder.

This also fixes the Github Pages deploy for this repo because
build-and-deploy.yml specifies the CLEAN=true option for the deploy
action. This means that it will perform "git clean" before publishing
contents. Because dist/bundle.js and dist/bundle.js.map were tracked
by git, git was discarding the contents of the generated files.